Read requirements.txt file to fillup install_requires in setup.py
Currently, install_requires field in setup function in setup.py is
hardcoded for the versions of packages required. Inorder to change the
version of a package, we need to change the package version in
requirements.txt and setup.py which is unecessary work. Read
requirements.txt directly for the required packages.

I encountered this when I was trying to reproduce why prompt_toolkit v1
was failing for pytify.